Added tests for apache domains.
Included a utility method for combining two dictionaries to assist in multiple dictionaries of subdomains (one common and another specific to the domain).

domain method is used to iterate through the subdomains prepending them to the domain and making a call via the apache class to determine the liveness of the domain. The result is saved to a dictionary using the fully qualified domain name as a key.

These classes need massive refactoring to bring the domain and combineDictionaries methods into the apache class.
Unit tests also need to be written for the managing of the domain merging.